Fix bug that occurs when creating a new Item
Bug fix for https://github.com/ArasLabs/style-non-current-items/issues/7
It uses the created_on property to avoid that this Method is triggered on a new created item. 
This fix is based on an idea provided by another Aras user. Unfortunately I wasn´t able to determine the originator of this workaround,